requirements.txt
openai==1.3.0
python-dotenv==1.0.0
colorama==0.4.6
rich==13.6.0
prompt-toolkit==3.0.39
pexpect==4.8.0; sys_platform != 'win32'
pyreadline3==3.4.1; sys_platform == 'win32'
# Browser automation dependencies
selenium==4.18.1
webdriver-manager==4.0.1
playwright==1.42.0 